Fixed various bugs, added statistical grade analysis
Refactored some of the controller code, got rid of the oscar api for
just /:course/:year.  This was done as this guesses the current
semester, but /:course/:year/:semester wouldn't really work without
some crazy logic…therefore just decided to streamline it all.  Maybe in
the future I'll just list the semesters for that api call…who knows.

I've also added a statistics generator so that after the grade dataset
has been updated, a call to the updateStatistics function in the grade
api will literally update all statistics of the dataset.  This uses
jStat.