Fixed search results
was using scrolls, but it wasn't the best way.  Instead I've recently
discovered the 'from' tag in the query source.  So I'm now adding a get
parameter of ?from=(which result to fetch).  I'm also defaulting to 25
results per search query.

the query function now takes in a param obj as it was getting pretty
rediculous to keep track of all the calls….now that I think of it, I
should probably do the same for oscar_api (will happen later this week).